changed type of matrix row, col and nonzero from int to long

+++ sparse_formats.h
@@ -35,7 +35,7 @@ typedef struct csr_matrix
 {//if ith row is empty Ap[i] = Ap[i+1]
     unsigned int index_type;
     float value_type;
-    unsigned int num_rows, num_cols, num_nonzeros,density_ppm;
+    unsigned long num_rows, num_cols, num_nonzeros,density_ppm;
     double density_perc,nz_per_row,stddev;
 
     unsigned int * Ap;  //row pointer
